Use vscode-languageclient as extension basis
This change migrates the extension codebase to be based on the
vscode-languageclient npm package.  This package provides a standard
language service client which speaks an updated protocol and is able to
feed all VS Code language features based on interactions with the language
server.  Many unnecessary files have now been deleted as a result of this
change.