Update package.json
This change updates package.json to add the "Debuggers" category and also
to change the command that is run for the PowerShell debug adapter.  Since
there isn't a good way to send a command line parameter to the debug
adapter (to switch to debug adapter mode) I've put the proper invocation
into a .cmd file which gets run instead.