Add choice prompt handling support
This change adds choice prompt handling support by responding to the new
powerShell/showChoicePrompt message.  When received, the extension will
show a quick pick list with the possible choices.  If the user selects an
option with 'enter' it will be sent for processing.  If the quick pick
gets cancelled either by pressing `ESC` or the UI losing focus, the prompt
will be cancelled in PowerShell.