Turn on verbose logging for debug adapter
This change turns on verbose logging for the debug adapter by default.
This is helpful in the short term because it is difficult for users to go
enable this setting (they have to edit package.json in their extension
folder).  In the future we will find a better way to configure this
setting for the debug adapter.