Initial attempt to fix #122 analyzer settings path
This adds a scriptAnalysis.settingsPath setting but uncovers some issues about how to be an absolute path back to the language host from the client.  Depending on where the user "set" this setting a relative path has a different meaning.  If set in .vscode\settings.json then relative probably should be relative to the workspaceRoot.  In the global settings file, it should probably be the extensionInstallDir.  If in the user's settings file, then what ... their home dir?

If we treated this like the other HostPaths then relative would *always* be relative to the extensionInstallDir.  If we could get VSCode to expand ${workspaceRoot} that would allow the user to specify a path in the current workspace folder.