Add support for loading host profile scripts
This change adds support for loading both host-specific and host-agnostic
profile scripts for both the current user and all users on the same
system.  For VS Code, the profile names are Microsoft.VSCode_profile.ps1
and profile.ps1.  The new 'enableProfileLoading' setting controls whether
profiles are loaded automatically when the language server starts up.  For
now this setting is turned off by default.

Resolves #124.