Consume Editor Services host as PowerShell module
This change refactors the existing PowerShell Editor Services host
management and editor client code to use the PSES host as a PowerShell
module rather than its previous executable form.  This involves launching
the powershell.exe process and loading up the PSES host as a module in
that process.  It also involves communicating with the PSES host via named
pipes instead of standard in/out as we did previously.

This change can be understood further by looking at the following pull
request at the PowerShellEditorServices repo:

https://github.com/PowerShell/PowerShellEditorServices/pull/256