Fix #182: Run Selection doesn't capture whole line
This change fixes an issue with the "Run Selection" command when it is
executed on a line without any text being selected.  In this case the
expected behavior is to execute the entire line but only the line contents
up to the cursor position were being executed.  The fix is to change the
content selection code to ensure that the line's full content is executed.