Replace program with script in launch.json.
"program" is left in place with updated description indicating the user should use the  "script" property instead.  This will allow existing PowerShell projects to continue work with newer versions of PSES.

Any new launch.json files created with this updated extension will generate just "script". This means that if someone loads the project in an older version of the extension, it won't launch the debugger correctly.  The question is, given how *easy* it is to update PSES, do we think this will be a big problem?

I don't "think" so and I think the term "script" will make more sense to PowerShell users in the long run.  That said, I'm totally OK with rejecting this PR if other folks do not agree.  BTW there is a companion PR on PSES to implement the PSES side of this change.