Fix #292: Check for Homebrew's OpenSSL libraries correctly
This change improves the way we check for OpenSSL libraries that must be
installed on the user's macOS system before PowerShell Editor Services can
be loaded successfully.  Previously we were looking for a path that
required a symbolic link to be created after installing OpenSSL.  It turns
out that PowerShell on macOS has built-in library paths which target
Homebrew's installation path for OpenSSL.  We now check primarily for this
path before looking at the system-wide installation path.