Change order of BundledModulesPath in PSModulePath
This change modifies the order of BundledModulesPath in our manipulation
of the PSModulePath to include modules with the PowerShell extension.  To
ensure that newer modules on the user's system are used for certain
integrations (like PSScriptAnalyzer or Plaster), the BundledModulesPath
needs to go at the end of PSModulePath, not the beginning.