Fix #396: Extension commands loaded in profile are not being registered
This change fixes an issue caused by the recent extension code refactoring
which prevents extension commands from being registered in the extension
when loaded in the user's profile.  This was caused by the LanguageClient
not being given to the ExtensionCommandsFeature early enough for it to
register its handler for the extensionCommandAdded event from the language
server.