Add new "attach" and "interactive" debugging configurations
This change adds new configuration possibilies for PowerShell debugging in
the extension.  The first is simply to allow the user to not specify a
script file in their "launch" configuration; when such a configuration is
launched it merely drops them into an interactive shell in the Debug
Console where they can enter whatever commands they like.

The other new configuration is a new "attach" configuration which allows
the user to specify a process ID or a remote computer and process ID to
which the debugger should be attached when it launches.  This provides a
smooth attach/detach experience so that the user does not need to go into
an interactive console to initiate the Enter-PSSession and
Enter-PSHostProcess commands manually (though this is still possible and
supported in interactive debugging mode).