Add handler for editor/closeFile request
This change adds a handler for the editor/closeFile request which is part
of the editor extensibility model of PowerShell Editor Services.  This
request currently gets fired when the user exits a remote debugging
session where a remote file had been opened.  It will later be exposed
through the $psEditor API.