Fix #418: Debugger should retry if extension isn't loaded yet
This change adds retry logic to debugAdapter.ts so that it will check for
the language server's session file once a second for 30 seconds until it
appears.  If the file appears debugging will resume, otherwise a specific
error message will be displayed saying that the debugger timed out waiting
for the extension to initialize.

This is a temporary fix until we get to leverage VS Code 1.9.0's new
debugger launch capabilities.