Add startSessionCommand handler for debugger configuration
This change leverages the new "startSessionCommand" configuration parameter
for debugger contributions which allows us to provide a command that will
take a debug launch configuration, tweak it, and then launch the debugger
ourselves.  This allows us to do any necessary launching or
pre-configuration of our debug adapter.  It also allows us to create a
default launch configuration on the fly if one doesn't already exist.
This means that users can now debug the current script without a
launch.json either with or without a folder open!