Add "remote files" feature that handles files from remote sessions
This change adds a new remote files feature which basically just handles
the cleanup of remote session files when the extension is loaded or
unloaded.  This is to ensure that the user doesn't accidentally edit a
locally cached remote file thinking it is their local copy.  It also shows
a warning to the user when they save changes to a remote file because we
currently don't propagate those changes back to the remote session.