Don't send empty environment vars when not Windows PowerShell dev build
This change fixes an issue where $env:PROCESSOR_ARCHITECTURE is not being
set properly in the language server's powershell.exe process which caused
an exception when gathering that variable.  The problem was caused by the
environment variable setting behavior when the Windows PowerShell
developer build setting is set to false.  In this case, an empty
environment variable object was being passed to the process, effectively
clearing its environment variables.

This is one half of the fix to PowerShell/PowerShellEditorServices#346.