Apply bundledModulesPath setting only when in development mode
This change improves our handling of the bundledModulesPath setting by
only making it active when VS Code is in development mode (either running
in the debugger or with the --extensionDevelopmentPath setting).  This
enables extension developers to use their installed PowerShell extension's
bundled modules in normal circumstances, only switching to the
bundledModulesPath when they are running a development build.

This code also sets the default bundledModulesPath to a
peer (co-located) PowerShellEditorServices folder since most developers
will probably clone the two repos into the same folder.  One would only
need to set their bundledModulesPath if their PSES repo folder was in an
unconventional path.

The bundledModulesPath setting will likely be renamed to something more
appropriate in the next update.