Add a PowerShell build script and wire it up in tasks.json
This change introduces a new vscode-powershell.build.ps1 script powered by
Invoke-Build.  The goal of this script is to handle build tasks which
would affect both the PowerShell extension and PowerShell Editor Services
so that it's much easier to hit F5 and build everything before launching
the debugger.

This is also a starting point for release automation so we can start
shipping faster!