Prevent concurrent formatting requests for a document
A user can execute `Format Document` command on a document even if the same document is already being formatted. This sends another, redundant, formatting request to the server for the same file. Another annoying side effect is that a new formatting status for the same file gets created on the status bar.