Find module display quick pick list first.
This PR puts the quick pick list up first with "Cancel" as the only item and text that indicates what is happening and that it could take some time.

@daviwil - please check my use of the CancellationTokenSource - especially in the pickPowerShellModule function where the request completes but the user (or timeout) has already cancelled the op.  Not sure about returning Promise.resolve("").   My JavaScript FU is pretty weak.  Also feel free to help with the wording of the various messages in the UI.

Finally what timeout should we use? I set it for 1 minute but that might be long.   It seems to take ~20 seconds on my machine.  But as the gallery grows, so could that time.

Finally, for the future, it would be nice if we could detect a user key press and if the find isn't done - cancel and restart a new find based on what the user typed.  Also, need timer support so that you only kick off a new find after say 1 second of paused typing.