Add params to DebugTest.ps1. (#485)
* Add params to DebugTest.ps1.

This will align with part 2 of my debugging blog post for Scripting Guys.  This allows testing of launch  config that passes arguments and makes it easier to demo "attach to" but setting the delay a bit longer (say 2 secs).  Also, updated  launch.json to start with the default configuration before add tle Launch Script Configuration.

* Removed line that was used for testing.