Fix #401: Unsupported PowerShell version should display an error
This change improves a previous fix where we report an unsupported
PowerShell version when the user attempts to start the extension.  The
previous fix utilized ConvertTo-Json to write out the session details;
this caused a failure since this cmdlet was introduced in PowerShell v3.
We now write out teh JSON manually in this case.