Fix #532: DEVPATH env variable not being set for integrated console
This change adds a workaround for setting the DEVPATH environment variable
when launching the language server inside of the integrated terminal UI.
This environment variable needs to be set for Windows PowerShell
development builds but it currently can't because there's no API parameter
for passing along environment variables.  The workaround is to generate a
batch script which can set the environment variable before launching
PowerShell.