Add setting to determine console focus behavior on script execution
This change adds a new setting that controls whether focus changes to the
integrated console when scripts are executed or debugged:

- `powershell.integratedConsole.focusConsoleOnExecute`

When true, it causes the focus to get set to the console on any script
execution.  It is set to true by default to ensure that screen readers
pick up the execution output once it starts.

Resolves #588.