Change alignAssignmentsInHashtable to alignPropertyValuePairs
The property value pairs defined in a DSC configuration are in fact key value pairs of a hashtable. This can be easily verified by parsing a DSC configuration and examining the AST nodes. However, while naming they are referred to as part of DSC configuration and not a hashtable. Therefore the option name "alignAssignmentsInHashtable" would sound somewhat misleading to users, if enabling the option would format even the property value pairs in a DSC configuration.