Add Install-VSCode.ps1 script
This change adds a new script that can be used to install Visual Studio
Code and the PowerShell extension.  The intent is for this script to be
shipped to the PowerShell Gallery so that one can easily get the script
and install VS Code in a single line of PowerShell.