Restarting PowerShell session now retains previous configuration
This change fixes #807 which states that restarting the current
PowerShell session using the "Restart Current Session" command doesn't
retain the current session configuration if the user previously switched
their configuration using the session menu.  The fix is to change the
call to restartSession to use the current session configuration when
this command is invoked.