Add CodeLens support for function and cmdlet references
This change adds support for CodeLenses that display the number of
references of a given function or cmdlet defined in a script file.  The
actual implementation of this behavior is on the server side but we had
to write a LanguageClient middleware to translate the arguments of the
references command when it was sent from the server.

Resolves #507
Resolves #827 (unrelated to CodeLenses but related to LanguageClient
output)