Language service crashes shouldn't keep process and terminal alive
This change updates the Start-EditorServices.ps1 script to not hold the
process open when the language service crashes.  This forced delay was
keeping the user from receiving the appropriate restart message so that
they can restart their session and continue with their work.  The fix is
to remove this delay so that the process exits immediately allowing the
extension to show the restart prompt.