Simplify and persist PowerShell session configuration changes
This change improves our session menu so that when the user selects a
different PowerShell version the choice is persisted in their user
settings.  The previous code for managing session configurations has
been mostly removed because it was unnecessarily complex.

We've also added a new "powershell.powerShellExePath" setting to replace
the now deprecated "powershell.developer.powerShellExePath" setting
since this is now a user-level setting.  If a user has the deprecated
setting in their configuration file, we will notify them and fix it
automatically if they wish.