Replace hardcoded c:\windows with ${env:windir}.
We used to have ${env.windir} but then VSCode broke us so we switched back to C:\Windows.

It turns out that the VSCode folks changed the syntax to ${env:windir} - replacing the dot with a colon.

Here's the VSCode issue describing this: https://github.com/Microsoft/vscode/issues/28365#issuecomment-308231940