Launch powershell.exe with -ExecutionPolicy Bypass
This change causes our PowerShell sessions to be launched with an
ExecutionPolicy of Bypass to fix some of the extension loading
failures that users have been reporting.  This setting avoids the prompt
which asks the user to accept running the signed
Start-EditorServices.ps1 script before the extension can be loaded.