Eliminate Pester duplicate failures in Problems view.
When tasks is set to 2.0.0, the tasks run in the terminal and in this case, Pester outputs a bunch of progress text to the terminal.  This results in duplicate failures being found by the problem matcher.

Work around this problem for now by eliminating the Pester progress output.  For more details see - https://github.com/Microsoft/vscode/issues/22065#issuecomment-313418250