Fix #1016: PowerShell extension doesn't load due to extension conflict
This change fixes an issue that was illuminated by the AzureRM Tools
extension where the `fs.appendFile` method provided by Node.js was not
being called with the correct number of parameters.

The AzureRM Tools extension loads up the AppInsights SDK for Node.js
which in turn loads a module called Zone.js.  Zone.js replaces built-in
Node.js modules like `fs` to instrument their function calls.  The
instrumented version of that module was not as permissive on our missing
`appendFile` parameter so it threw an exception which ultimately
prevented the PowerShell extension from loading.

This issue was fixed by populating the `callback` parameter of the
`appendFile` function so that the instrumented wrapper operates
correctly.