Update to use DebugConfigurationProvider
Fix #1046
Fix warning in status bar about ThemeColor

This fix requires PSES PR 556 to eliminate one PSES cause of the debug adapter to crash.

Also npm generated a package-lock.json file which, from what I've read, should be checked in.  I haven't included it in this PR - yet.  Thoughts?

Also, I updated min VSCode version to 1.17.0.  I'm not sure when the new DebugConfigurationProvider API was intro'd.  I  have tested with 1.17.2 and it works there so I'm reasonably confident it will work in 1.17.0.  We might be able to reach back a bit further.  We'd have to find out when this API was intro'd.