(maint) Modify tslint configuration for test files
Previously tslint was raising errors in Travis CI saying that the excluded test
fixtures directory was not included in the project.  This was by design however
it appears to be a known bug https://github.com/palantir/tslint/issues/3793.

This commit removes the exclude for test files from linting and adds a tslint
directive in the default index.ts file.  A tslint directive is used instead of
solving the issue because this is the default testing file for VS Code extesions
and shouldn't really be modified unless absolutely necessary.  In this instance
it was safer for a tslint directive.