Allow debugging in interactive session with no dir change (#1397)
Fix #1330

This PR depends on a corresponding PR to PSES to have it handle
null/empty string differently in the non-temp console case.

For the generateLaunchConfig case, we now pass "" as cwd to PSES.
That tells PSES to not change the directory *if* we aren't running
in a temp console.  If we are in a temp console, then use old logic to
set working dir.

Update "PowerShell Interactive Session" debug config to tell PSES
to not change the working dir.

Remove "program" field from launch config. This has been marked
deprecated for over a year now.

Change refs in ${workspaceRoot} to ${workpaceFolder} to work
better in a multi-workspace environment.

Remove unused imports/field in DebugSession.ts.