(GH-1428) Make region folding case insensitive and strict whitespace (#1430)
Previously the code folding feature would fail to find region blocks starting
with `Region` or ending with `EndRegion`.  This was due to the regex only
looking for lower case.  This commit updates the start and end region
detection to be case insensitive, similar to that defined in
https://github.com/Microsoft/vscode/commit/64186b0a262f0ff89a060cf8dbbf8de7ff831a00

This commit adds has strict whitespace rules for the region/endregion keyword
which means only `#region` will match whereas `# region` will not.

Previously the code folding had trouble detecting the beginning of comment
blocks and regions which were not indented.  This was due to the empty line
regex expecting at least one whitespace character to be considered "an empty
line".  This commit modifies the empty line regex to also allow an empty string
which is indeed, an empty line.