(GH-1459) Folding provider should not crash with mismatched region tokens (#1461)
Previously the folding provider would crash with an error if the document
contained mismatched begin and end region comments e.g. If the document
started with `# endregion`.  This was due to the token stack code always
assuming there was at least one element in the stack.  This commit modifies
the end region detection to only trigger if there was a previous begin region.