(GH-1458) Remove region folding from non-region areas (#1467)
Previously all of the folding ranges were either Region or Comment, however this
is confusing e.g. a Here String is neither a region or a comment.  The VS Code
API does state that the range kind property is optional.  This commit changes
the range kind for braces, parentheses and here strings to be null, that is,
neither a region or comment range. This makes editor commands like
'Fold All Regions' behave as a user expects.