(maint) Add tests to PR Checklist
While the project has only a handful of tests, it would be in the interest of
contributors to at least be prompted to think about adding tests for new
functionality.  This commit updates the PR template and adds a testing checkbox.
Note that these are not mandatory and can even be marked as N/A when people
raise a PR.