Change debugger type field back to "PowerShell" from poweshell
This allows the debugger to start but not sure why powershell doesn't
work. That said, the previous change meant that users would have to
update their launch.json to change the type to "powershell".  But
even when I do that, the debugger still doesn't launch. So I'm a bit
puzzled.  But this gets the debugger working again on master.

Also fixed some warnings in session.ts.