Update PowerShell.json
Added a basic parameter block (At line 574) that includes the cmdletbinding attribute.

As a user I would like a snippet to make it easier to add a param() section with one parameter declaration to get it started.  
The cursor is deposited to the right of the parameter variable.  If the user would like to add more parameters, they can type the comma, and proceed to use the "param" snippet to add another parameter.

This addresses issue #2080