Provide Run button in Editor Title Menu
This commit adds a editor title menu button for the
`powershell.Run` command which maps to the builtin VS Code Debug start
command. It places it in the primary navigation group, and
the uses `run` icon from the https://github.com/microsoft/vscode-icons
repo. This shows the button in the visible part of the menu when a
PowerShell file is open.