Bump CI images
We should considering setting up a matrix so we can purposefully test on
the older versions of these images. However, we at least need to bump
the macOS agent for Node, as this updated dependency requires at least
Node 12. The 10.14 image comes with Node 8, and the 10.15 image comes
with Node 14.