Fix release build pipeline
A backlog task of removing the `findPsesBuild` script and associated logic
caused us to end up with a release with the wrong artifacts. The tagged PSES
build was partially succeeded, which the download task then ignored as the
trigger, and instead of failing, fell back to this script which gave it the
latest fully succeeded build (an old version of PSES).

That PSES build was only partially successful because the "Install PowerShell
Daily" step failed due to a transient network error, and so this was
intentionally ignored as it did not affect the build. Or so we thought.