Export `InvokePesterStub.ps1` to `PowerShellEditorServices`
This fixes the path to the script which is now part of the
`PowerShellEditorServices` repository. This was necessary in order to sign it,
as it must be signed _before_ the VSIX is created. Moving it to the other repo
was the best way to solve the "chicken-egg" problem that presented.