Fix side effect logic in `ReleaseTools` module
This uses a new function `Use-Repository` that executes the given script
with the given repository as the current working directory, replacing
all the previously side effect driven logic, or at least, making it more
explicit. This was frankly easier than dealing with the Azure Pipeline
that clones the repository under a different name (specifically `s` if
its the only repo, that is, it's not named `vscode-powershell` in the
release stage, so the module broke).