Make CI build template actually generic
This now uses a repository resource or the triggering pipeline artifact
to provide the PowerShellEditorServices sources (or bits). Coupled with
the change to the build script, this means the template can be used to
build for CI or for a release.

We also moved to zipped pipeline artifacts as build artifacts have been
deprecated, but we need to reduce the cost of the stored artifacts
themselves. These are provided for developers to test and for the
release pipeline to use.