Pipeline-ify release process
This sets up the Azure DevOps release pipeline to not only build, test,
and sign the bits for a release (and to do so using signed
PowerShellEditorServices bits), but to automatically create the GitHub
draft release where it directly uploads the artifacts. After manual
approval, it will automatically publish the extension and installation
script to their respective registries (the Visual Studio Code
marketplace and the PowerShell Gallery). Best of all, this process is
kicked off automatically after a successful release build of
PowerShellEditorServices.