Remove `*All` build tasks from project
These existed to invoke tasks in the associated PowerShellEditorServices
folder; however, the `Build` task was previously fixed to automatically
build and copy that dependency when needed. Furthermore, the existing VS
Code tasks meant that for developers' ease-of-use we could already
clean, build, and test that dependency. Finally, CI did not depend on
the `All` section of these for testing, as we have separate CI/CD
pipelines for each project, so these were redundant and confusing.