Setup `nyc` to produce code coverage reports
This extends the public TypeScript configuration and sets it to check
just our source files. The `all` setting must be `false` otherwise the
tool crashes with an OOM because the test runner launches VS Code
instances. The `.vscode-test` folder must be excluded or `nyc` will
attempt to instrument Code's own sources. The `out` and `src` folders
must both be included so that the TypeScript source mapping works (along
with the `require` settings in the configuration and in Mocha). The
coverage check is set to false because we'll enable failures for reduced
coverage later. The Cobertura report is for Azure DevOps (cobertura) and
a summary is printed to the console.