Fix `EnableISEMode` for Azure Data Studio (#3735)
Because ADS stripped out VS Code's debugger, the `debug` settings don't
exist, causing our `update()` call for the debug setting to throw an
error to the user when they run `EnableISEMode`. I couldn't find a
reliable way to test if the extension is running in the ADS fork instead
of the upstream VS Code, but a simple `try {} catch {}` that allows the
funciton to continue is a reliable fix, as we can't do anything about
this setting regardless.