Add magic marker to `specifyScriptArgs` to disable argument escaping
Coupled with a change to PSES, this enables us to workaround VS Code's
API limitations where our interactive "get user's arguments" command can
only return a string, not an array of strings, which is then passed as a
single argument to PSES. Because it likely contains spaces (and multiple
arguments), we need to instruct PSES to not quote the string in this
special case. It would be easier if we could just pass a list of string
arguments, but so it goes.