Use new `isTransient` API to prevent duplicate integrated consoles
This new API available today in VS Code 1.65.0 prevents Code from saving
and reloading the integrated console when the extension is reloaded.
This is necessary because the extension always spawns a new terminal to
host PowerShell Editor Services when it starts, and Code's newish
feature to save and restore terminals caused these to be duplicated.
When we update the VS Code engine, we can remove the cast to `any`.